Fix several corner cases in delayMicroseconds()
The formulas in delayMicroseconds for odd frequencies use the expression
us = (us * 30199) >> 16 or us = (us * 60398) >> 16.  These reduce us,
which must be taken into account when subtracting from the result.
If the subtraction rolls under, we get a delay that is 1000x too long.

This was the case for 11.0592 MHz at 5us, 3.6864 MHz at 13 and 14us,
and for 1.8432 MHz at 13us and 26..54 us.  Some other frequencies
had the rollunder case coded, but never entered it due to plain luck.